refactor: moved WasmBucketMount into its own file
